# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Research Lab API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RepositoryInput {
  /** GitHub repository URL */
  url: string;
}

export interface Repository {
  id: string;
  url: string;
  name: string;
  owner: string;
  language: string;
  framework: string;
  stars: number;
  forks: number;
  purpose: string;
  architectureSummary: string;
  /** Score 0-100 */
  maintainabilityScore: number;
  /** Score 0-100 */
  complexityScore: number;
  analyzedAt: string;
}

export type AgentOutputStatus =
  (typeof AgentOutputStatus)[keyof typeof AgentOutputStatus];

export const AgentOutputStatus = {
  complete: "complete",
  running: "running",
  pending: "pending",
} as const;

/**
 * @nullable
 */
export type AgentOutputRiskLevel =
  | (typeof AgentOutputRiskLevel)[keyof typeof AgentOutputRiskLevel]
  | null;

export const AgentOutputRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface AgentOutput {
  agentType: string;
  agentName: string;
  status: AgentOutputStatus;
  summary: string;
  findings: string[];
  recommendations: string[];
  /** @nullable */
  riskLevel?: AgentOutputRiskLevel;
}

export type FolderNodeType =
  (typeof FolderNodeType)[keyof typeof FolderNodeType];

export const FolderNodeType = {
  file: "file",
  directory: "directory",
} as const;

export interface FolderNode {
  name: string;
  type: FolderNodeType;
  children?: FolderNode[];
}

export interface Dependency {
  name: string;
  version: string;
  purpose: string;
}

export interface ArchitectureMap {
  folderTree: FolderNode[];
  keyFiles: string[];
  dependencies: Dependency[];
  componentRelationships: string[];
  dataFlow: string[];
}

export type IssuePriority = (typeof IssuePriority)[keyof typeof IssuePriority];

export const IssuePriority = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export type IssueType = (typeof IssueType)[keyof typeof IssueType];

export const IssueType = {
  bug: "bug",
  performance: "performance",
  security: "security",
  refactor: "refactor",
  documentation: "documentation",
  enhancement: "enhancement",
} as const;

export interface Issue {
  id: string;
  title: string;
  priority: IssuePriority;
  type: IssueType;
  description: string;
  suggestedFix: string;
  affectedFiles: string[];
}

export interface PrSummary {
  title: string;
  summary: string;
  implementationNotes: string[];
  testingChecklist: string[];
  riskNotes: string[];
}

export interface HealthDimension {
  id: string;
  label: string;
  score: number;
  description: string;
  insight: string;
}

export interface HealthScore {
  overallScore: number;
  grade: string;
  dimensions: HealthDimension[];
  summary: string;
}

export type DependencyItemRiskLevel =
  (typeof DependencyItemRiskLevel)[keyof typeof DependencyItemRiskLevel];

export const DependencyItemRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface DependencyItem {
  name: string;
  category: string;
  riskLevel: DependencyItemRiskLevel;
  reason: string;
}

export type DependencyRiskOverallRisk =
  (typeof DependencyRiskOverallRisk)[keyof typeof DependencyRiskOverallRisk];

export const DependencyRiskOverallRisk = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface DependencyRisk {
  totalCount: number;
  majorFrameworks: string[];
  estimatedOutdated: number;
  overallRisk: DependencyRiskOverallRisk;
  bundleComplexity: string;
  warnings: string[];
  recommendations: string[];
  items: DependencyItem[];
}

export type TimelinePhaseStatus =
  (typeof TimelinePhaseStatus)[keyof typeof TimelinePhaseStatus];

export const TimelinePhaseStatus = {
  complete: "complete",
  "in-progress": "in-progress",
  upcoming: "upcoming",
} as const;

export interface TimelinePhase {
  phase: string;
  phaseNumber: number;
  weeks: number;
  milestone: string;
  tasks: string[];
  status: TimelinePhaseStatus;
}

export interface EngineeringTimeline {
  phases: TimelinePhase[];
  totalWeeks: number;
  deploymentReadiness: string;
  projectType: string;
  summary: string;
}

export type ArchVizNodeLayer =
  (typeof ArchVizNodeLayer)[keyof typeof ArchVizNodeLayer];

export const ArchVizNodeLayer = {
  frontend: "frontend",
  backend: "backend",
  api: "api",
  services: "services",
  database: "database",
  config: "config",
  testing: "testing",
  shared: "shared",
} as const;

export interface ArchVizNode {
  id: string;
  label: string;
  layer: ArchVizNodeLayer;
  type: string;
  description?: string;
  files?: string[];
}

export interface ArchVizConnection {
  from: string;
  to: string;
  label: string;
}

export interface ArchViz {
  nodes: ArchVizNode[];
  connections: ArchVizConnection[];
  projectType: string;
  layers: string[];
}

export interface OnboardingStep {
  step: number;
  title: string;
  command: string;
  description: string;
}

export type BeginnerIssueDifficulty =
  (typeof BeginnerIssueDifficulty)[keyof typeof BeginnerIssueDifficulty];

export const BeginnerIssueDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
} as const;

export interface BeginnerIssue {
  title: string;
  difficulty: BeginnerIssueDifficulty;
  area: string;
  description: string;
}

export interface OnboardingGuide {
  setupSteps: OnboardingStep[];
  firstFilesToRead: string[];
  beginnerIssues: BeginnerIssue[];
  architectureOverview: string;
  localDevSteps: string[];
  packageManager: string;
  estimatedSetupMinutes: number;
}