# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Research Lab API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: repositories
    description: Repository analysis operations
  - name: agents
    description: AI agent operations
  - name: architecture
    description: Architecture map operations
  - name: issues
    description: Issue generation operations
  - name: prs
    description: PR summary operations
  - name: enhancements
    description: Enhanced analysis features

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /repositories/analyze:
    post:
      operationId: analyzeRepository
      tags: [repositories]
      summary: Analyze a GitHub repository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepositoryInput"
      responses:
        "200":
          description: Repository analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"

  /repositories/{repoId}:
    get:
      operationId: getRepository
      tags: [repositories]
      summary: Get repository analysis by ID
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Repository details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
        "404":
          description: Not found

  /repositories/{repoId}/agents:
    get:
      operationId: getAgents
      tags: [agents]
      summary: Get all AI agent outputs for a repository
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of agent outputs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentOutput"

  /repositories/{repoId}/agents/{agentType}:
    get:
      operationId: getAgentOutput
      tags: [agents]
      summary: Get a specific AI agent output
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: string
        - name: agentType
          in: path
          required: true
          schema:
            type: string
            enum:
              - research-lead
              - architecture
              - code-review
              - issue-triage
              - fix-engineer
              - pr-writer
      responses:
        "200":
          description: Agent output
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentOutput"

  /repositories/{repoId}/architecture:
    get:
      operationId: getArchitecture
      tags: [architecture]
      summary: Get architecture map for a repository
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Architecture map
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchitectureMap"

  /repositories/{repoId}/issues:
    get:
      operationId: getIssues
      tags: [issues]
      summary: Get generated issues for a repository
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Issue"

  /repositories/{repoId}/pr-summary:
    get:
      operationId: getPrSummary
      tags: [prs]
      summary: Get PR summary for a repository
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PR summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrSummary"

  /repositories/{repoId}/health-score:
    get:
      operationId: getHealthScore
      tags: [enhancements]
      summary: Get repository health scores across 6 dimensions
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Health score report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthScore"
        "404":
          description: Not found

  /repositories/{repoId}/dependency-risk:
    get:
      operationId: getDependencyRisk
      tags: [enhancements]
      summary: Analyze dependency risk for a repository
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dependency risk report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DependencyRisk"
        "404":
          description: Not found

  /repositories/{repoId}/timeline:
    get:
      operationId: getTimeline
      tags: [enhancements]
      summary: Get engineering timeline roadmap for a repository
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Engineering timeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EngineeringTimeline"
        "404":
          description: Not found

  /repositories/{repoId}/arch-viz:
    get:
      operationId: getArchViz
      tags: [enhancements]
      summary: Get architecture visualization data
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Architecture visualization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchViz"
        "404":
          description: Not found

  /repositories/{repoId}/onboarding:
    get:
      operationId: getOnboarding
      tags: [enhancements]
      summary: Get contributor onboarding guide
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Onboarding guide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingGuide"
        "404":
          description: Not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    RepositoryInput:
      type: object
      required: [url]
      properties:
        url:
          type: string
          description: GitHub repository URL

    Repository:
      type: object
      required: [id, url, name, owner, language, framework, stars, forks, purpose, architectureSummary, maintainabilityScore, complexityScore, analyzedAt]
      properties:
        id:
          type: string
        url:
          type: string
        name:
          type: string
        owner:
          type: string
        language:
          type: string
        framework:
          type: string
        stars:
          type: integer
        forks:
          type: integer
        purpose:
          type: string
        architectureSummary:
          type: string
        maintainabilityScore:
          type: integer
          description: Score 0-100
        complexityScore:
          type: integer
          description: Score 0-100
        analyzedAt:
          type: string

    AgentOutput:
      type: object
      required: [agentType, agentName, status, summary, findings, recommendations]
      properties:
        agentType:
          type: string
        agentName:
          type: string
        status:
          type: string
          enum: [complete, running, pending]
        summary:
          type: string
        findings:
          type: array
          items:
            type: string
        recommendations:
          type: array
          items:
            type: string
        riskLevel:
          type: ["string", "null"]
          enum: [low, medium, high, null]

    FolderNode:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [file, directory]
        children:
          type: array
          items:
            $ref: "#/components/schemas/FolderNode"

    Dependency:
      type: object
      required: [name, version, purpose]
      properties:
        name:
          type: string
        version:
          type: string
        purpose:
          type: string

    ArchitectureMap:
      type: object
      required: [folderTree, keyFiles, dependencies, componentRelationships, dataFlow]
      properties:
        folderTree:
          type: array
          items:
            $ref: "#/components/schemas/FolderNode"
        keyFiles:
          type: array
          items:
            type: string
        dependencies:
          type: array
          items:
            $ref: "#/components/schemas/Dependency"
        componentRelationships:
          type: array
          items:
            type: string
        dataFlow:
          type: array
          items:
            type: string

    Issue:
      type: object
      required: [id, title, priority, type, description, suggestedFix, affectedFiles]
      properties:
        id:
          type: string
        title:
          type: string
        priority:
          type: string
          enum: [critical, high, medium, low]
        type:
          type: string
          enum: [bug, performance, security, refactor, documentation, enhancement]
        description:
          type: string
        suggestedFix:
          type: string
        affectedFiles:
          type: array
          items:
            type: string

    PrSummary:
      type: object
      required: [title, summary, implementationNotes, testingChecklist, riskNotes]
      properties:
        title:
          type: string
        summary:
          type: string
        implementationNotes:
          type: array
          items:
            type: string
        testingChecklist:
          type: array
          items:
            type: string
        riskNotes:
          type: array
          items:
            type: string

    # ── Enhancement schemas ────────────────────────────────────────────────────

    HealthDimension:
      type: object
      required: [id, label, score, description, insight]
      properties:
        id:
          type: string
        label:
          type: string
        score:
          type: integer
        description:
          type: string
        insight:
          type: string

    HealthScore:
      type: object
      required: [overallScore, grade, dimensions, summary]
      properties:
        overallScore:
          type: integer
        grade:
          type: string
        dimensions:
          type: array
          items:
            $ref: "#/components/schemas/HealthDimension"
        summary:
          type: string

    DependencyItem:
      type: object
      required: [name, category, riskLevel, reason]
      properties:
        name:
          type: string
        category:
          type: string
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        reason:
          type: string

    DependencyRisk:
      type: object
      required: [totalCount, majorFrameworks, estimatedOutdated, overallRisk, bundleComplexity, warnings, recommendations, items]
      properties:
        totalCount:
          type: integer
        majorFrameworks:
          type: array
          items:
            type: string
        estimatedOutdated:
          type: integer
        overallRisk:
          type: string
          enum: [low, medium, high, critical]
        bundleComplexity:
          type: string
        warnings:
          type: array
          items:
            type: string
        recommendations:
          type: array
          items:
            type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/DependencyItem"

    TimelinePhase:
      type: object
      required: [phase, phaseNumber, weeks, milestone, tasks, status]
      properties:
        phase:
          type: string
        phaseNumber:
          type: integer
        weeks:
          type: integer
        milestone:
          type: string
        tasks:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [complete, in-progress, upcoming]

    EngineeringTimeline:
      type: object
      required: [phases, totalWeeks, deploymentReadiness, projectType, summary]
      properties:
        phases:
          type: array
          items:
            $ref: "#/components/schemas/TimelinePhase"
        totalWeeks:
          type: integer
        deploymentReadiness:
          type: string
        projectType:
          type: string
        summary:
          type: string

    ArchVizNode:
      type: object
      required: [id, label, layer, type]
      properties:
        id:
          type: string
        label:
          type: string
        layer:
          type: string
          enum: [frontend, backend, api, services, database, config, testing, shared]
        type:
          type: string
        description:
          type: string
        files:
          type: array
          items:
            type: string

    ArchVizConnection:
      type: object
      required: [from, to, label]
      properties:
        from:
          type: string
        to:
          type: string
        label:
          type: string

    ArchViz:
      type: object
      required: [nodes, connections, projectType, layers]
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/ArchVizNode"
        connections:
          type: array
          items:
            $ref: "#/components/schemas/ArchVizConnection"
        projectType:
          type: string
        layers:
          type: array
          items:
            type: string

    OnboardingStep:
      type: object
      required: [step, title, command, description]
      properties:
        step:
          type: integer
        title:
          type: string
        command:
          type: string
        description:
          type: string

    BeginnerIssue:
      type: object
      required: [title, difficulty, area, description]
      properties:
        title:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate]
        area:
          type: string
        description:
          type: string

    OnboardingGuide:
      type: object
      required: [setupSteps, firstFilesToRead, beginnerIssues, architectureOverview, localDevSteps, packageManager, estimatedSetupMinutes]
      properties:
        setupSteps:
          type: array
          items:
            $ref: "#/components/schemas/OnboardingStep"
        firstFilesToRead:
          type: array
          items:
            type: string
        beginnerIssues:
          type: array
          items:
            $ref: "#/components/schemas/BeginnerIssue"
        architectureOverview:
          type: string
        localDevSteps:
          type: array
          items:
            type: string
        packageManager:
          type: string
        estimatedSetupMinutes:
          type: integer