# lib/api-zod/src/generated/types/archViz.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Research Lab API
 * OpenAPI spec version: 0.1.0
 */
import type { ArchVizConnection } from "./archVizConnection";
import type { ArchVizNode } from "./archVizNode";

export interface ArchViz {
  nodes: ArchVizNode[];
  connections: ArchVizConnection[];
  projectType: string;
  layers: string[];
}