# lib/api-zod/src/generated/types/archVizConnection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Research Lab API
 * OpenAPI spec version: 0.1.0
 */

export interface ArchVizConnection {
  from: string;
  to: string;
  label: string;
}