# lib/api-zod/src/generated/types/archVizNode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Research Lab API
 * OpenAPI spec version: 0.1.0
 */
import type { ArchVizNodeLayer } from "./archVizNodeLayer";

export interface ArchVizNode {
  id: string;
  label: string;
  layer: ArchVizNodeLayer;
  type: string;
  description?: string;
  files?: string[];
}