# lib/api-zod/src/generated/types/beginnerIssue.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Research Lab API
 * OpenAPI spec version: 0.1.0
 */
import type { BeginnerIssueDifficulty } from "./beginnerIssueDifficulty";

export interface BeginnerIssue {
  title: string;
  difficulty: BeginnerIssueDifficulty;
  area: string;
  description: string;
}