# lib/api-zod/src/generated/types/dependencyItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Research Lab API
 * OpenAPI spec version: 0.1.0
 */
import type { DependencyItemRiskLevel } from "./dependencyItemRiskLevel";

export interface DependencyItem {
  name: string;
  category: string;
  riskLevel: DependencyItemRiskLevel;
  reason: string;
}