# lib/api-zod/src/generated/types/dependencyRisk.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Research Lab API
 * OpenAPI spec version: 0.1.0
 */
import type { DependencyItem } from "./dependencyItem";
import type { DependencyRiskOverallRisk } from "./dependencyRiskOverallRisk";

export interface DependencyRisk {
  totalCount: number;
  majorFrameworks: string[];
  estimatedOutdated: number;
  overallRisk: DependencyRiskOverallRisk;
  bundleComplexity: string;
  warnings: string[];
  recommendations: string[];
  items: DependencyItem[];
}