# lib/api-zod/src/generated/types/engineeringTimeline.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Research Lab API
 * OpenAPI spec version: 0.1.0
 */
import type { TimelinePhase } from "./timelinePhase";

export interface EngineeringTimeline {
  phases: TimelinePhase[];
  totalWeeks: number;
  deploymentReadiness: string;
  projectType: string;
  summary: string;
}