# lib/api-zod/src/generated/types/healthDimension.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Research Lab API
 * OpenAPI spec version: 0.1.0
 */

export interface HealthDimension {
  id: string;
  label: string;
  score: number;
  description: string;
  insight: string;
}