# lib/api-zod/src/generated/types/healthScore.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Research Lab API
 * OpenAPI spec version: 0.1.0
 */
import type { HealthDimension } from "./healthDimension";

export interface HealthScore {
  overallScore: number;
  grade: string;
  dimensions: HealthDimension[];
  summary: string;
}