# lib/api-zod/src/generated/types/onboardingGuide.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Research Lab API
 * OpenAPI spec version: 0.1.0
 */
import type { BeginnerIssue } from "./beginnerIssue";
import type { OnboardingStep } from "./onboardingStep";

export interface OnboardingGuide {
  setupSteps: OnboardingStep[];
  firstFilesToRead: string[];
  beginnerIssues: BeginnerIssue[];
  architectureOverview: string;
  localDevSteps: string[];
  packageManager: string;
  estimatedSetupMinutes: number;
}