# lib/api-zod/src/generated/types/onboardingStep.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Research Lab API
 * OpenAPI spec version: 0.1.0
 */

export interface OnboardingStep {
  step: number;
  title: string;
  command: string;
  description: string;
}