# lib/api-zod/src/generated/types/timelinePhase.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Research Lab API
 * OpenAPI spec version: 0.1.0
 */
import type { TimelinePhaseStatus } from "./timelinePhaseStatus";

export interface TimelinePhase {
  phase: string;
  phaseNumber: number;
  weeks: number;
  milestone: string;
  tasks: string[];
  status: TimelinePhaseStatus;
}